Admin , Super Admin , Suggestion Box, and Glossary Features added

+++ anuvaad-fe/anuvaad-webapp/src/ui/containers/web/UserGlossary/ViewUserGlossary.jsx
@@ -16,6 +16,13 @@ import ViewGlossary from '../../../../flux/actions/apis/user_glossary/fetch_user
 import Spinner from "../../../components/web/common/Spinner";
 import DeleteGlossary from '../../../../flux/actions/apis/user_glossary/delete_glossary';
 import Snackbar from "../../../components/web/common/Snackbar";
+import { Button } from "@material-ui/core";
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import DeleteTmx from "../../../../flux/actions/apis/tmx/tmxDelete";
 
 var delete_glossary = require("../../../../utils/deleteGlossary.operation");
 
@@ -40,13 +47,16 @@ class MyGlossary extends React.Component {
             message: "",
             variant: 'success',
             loadMsg: "",
-            rowsToDelete: []
+            rowsToDelete: [],
+            openConfirmDialog: false,
+            openDeleteSelectedGlossaryConfirmDialogue: false
         }
     }
     getUserGlossary = () => {
         const { APITransport } = this.props
-        let userID = JSON.parse(localStorage.getItem("userProfile")).userID
-        let apiObj = new ViewGlossary(userID)
+        let userID = JSON.parse(localStorage.getItem("userProfile")).userID;
+        let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        let apiObj = new ViewGlossary(userID, orgID);
         APITransport(apiObj)
     }
     componentDidMount() {
@@ -66,18 +76,23 @@ class MyGlossary extends React.Component {
         }
     }
 
-    makeDeleteGlossaryAPICall = (userId, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []) => {
+    makeDeleteGlossaryAPICall = (userId, orgID, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []) => {
+        console.log(`userId, src, tgt, locale, reverseLocale, context, bulkDelete = false, deletionArray = []`);
+        console.log(userId, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray);
         this.setState({ open: true, message: 'Glossary deletion in progress...', variant: 'info' })
-        let apiObj = new DeleteGlossary(userId, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray)
+        // let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        let apiObj = new DeleteGlossary(userId, orgID, src, tgt, locale, reverseLocale, context, bulkDelete, deletionArray);
+        let payloadObj =  apiObj.getBody();
+        payloadObj.orgID && delete payloadObj.orgID;
         fetch(apiObj.apiEndPoint(), {
             method: 'post',
             headers: apiObj.getHeaders().headers,
-            body: JSON.stringify(apiObj.getBody())
+            body: JSON.stringify(payloadObj)
         })
             .then(async res => {
                 if (res.ok) {
                     this.setState({ open: false })
-                    let apiObj = new ViewGlossary(userId)
+                    let apiObj = new ViewGlossary(userId, orgID)
                     let { APITransport } = this.props
                     APITransport(apiObj)
                     return true;
@@ -88,9 +103,120 @@ class MyGlossary extends React.Component {
             })
     }
 
+    makeDeleteAllGlossaryAPICall = (userObj, context) => {
+        this.setState({ open: true, message: 'Glossary deletion in progress...', variant: 'info', openConfirmDialog: false })
+        let apiObj = new DeleteTmx(userObj, context)
+        // console.log("apiObj.getBody()", apiObj.getBody());
+        let userID = JSON.parse(localStorage.getItem("userProfile")).userID;
+        let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        fetch(apiObj.apiEndPoint(), {
+          method: 'post',
+          headers: apiObj.getHeaders().headers,
+          body: JSON.stringify(apiObj.getBody())
+        })
+          .then(async res => {
+            if (res.ok) {
+              this.setState({ open: false })
+              let apiObj = new ViewGlossary(userID, orgID)
+              let { APITransport } = this.props
+              APITransport(apiObj)
+              return true;
+            } else {
+              this.setState({ open: true, message: 'Glossary deletion failed', variant: 'error' })
+              return false;
+            }
+          })
+      }
+
+      handleDeleteAllGlossary = () => {
+        console.log("handleDeleteAllGlossary")
+        this.setState({openConfirmDialog : true})
+      }
+
+      handleCloseConfirmBox = () => {
+        this.setState({openConfirmDialog : false})
+      }
+
+      renderDeleteAllGlossaryButton = () => {
+        return (
+          <div style={{ textAlign: "end", marginBottom : "1rem" }}>
+            <Button
+              onClick={() => this.handleDeleteAllGlossary()}
+              variant="contained"
+                color="primary"
+                style={{
+                    borderRadius: "10px",
+                    color: "#FFFFFF",
+                    backgroundColor: "#1C9AB7",
+                    height: "35px",
+                    fontSize: "16px",
+                    textTransform: "none",
+                }}
+                size="small"
+            >
+              Delete All Glossary
+            </Button>
+            <Dialog
+              open={this.state.openConfirmDialog}
+              onClose={()=>this.handleCloseConfirmBox()}
+              aria-labelledby="alert-dialog-title"
+              aria-describedby="alert-dialog-description"
+            >
+              <DialogTitle id="alert-dialog-title">{"Delete all glossary"}</DialogTitle>
+              <DialogContent>
+                <DialogContentText id="alert-dialog-description">
+                  Are you sure you want to delete all glossary?
+                </DialogContentText>
+              </DialogContent>
+              <DialogActions>
+                <Button onClick={()=>this.makeDeleteAllGlossaryAPICall({userID : JSON.parse(localStorage.getItem("userProfile")).userID},"JUDICIARY")} color="primary">
+                  Confirm
+                </Button>
+                <Button onClick={()=>this.handleCloseConfirmBox()} color="primary" autoFocus>
+                  Cancel
+                </Button>
+              </DialogActions>
+            </Dialog>
+          </div>
+        )
+      }
+
+      handleDeleteSelectedGlossaryBox = () => {
+        this.setState({openDeleteSelectedGlossaryConfirmDialogue : false})
+      }
+
+      renderDeleteSelectedGlossaryConfirmBox = () => {
+        return (
+          <div style={{ textAlign: "end", marginBottom : "1rem" }}>
+            <Dialog
+              open={this.state.openDeleteSelectedGlossaryConfirmDialogue}
+              onClose={()=>this.handleDeleteSelectedGlossaryBox()}
+              aria-labelledby="alert-dialog-title"
+              aria-describedby="alert-dialog-description"
+            >
+              <DialogTitle id="alert-dialog-title">{"Delete Selected glossary"}</DialogTitle>
+              <DialogContent>
+                <DialogContentText id="alert-dialog-description">
+                  Are you sure you want to delete selected glossary?
+                </DialogContentText>
+              </DialogContent>
+              <DialogActions>
+                <Button onClick={()=>this.deleteMultipleRows()} color="primary">
+                  Confirm
+                </Button>
+                <Button onClick={()=>this.handleDeleteSelectedGlossaryBox()} color="primary" autoFocus>
+                  Cancel
+                </Button>
+              </DialogActions>
+            </Dialog>
+          </div>
+        )
+      }
+
     handleDeleteGlossary = (dataArray) => {
         let reverseLocale = dataArray[3].split("|").reverse().join("|");
-        this.makeDeleteGlossaryAPICall(dataArray[2], dataArray[0], dataArray[1], dataArray[3], reverseLocale, dataArray[4])
+        let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+        this.makeDeleteGlossaryAPICall(dataArray[2], orgID, dataArray[0], dataArray[1], dataArray[3], reverseLocale, dataArray[4])
     }
 
     handleClose = () => {
@@ -99,10 +225,12 @@ class MyGlossary extends React.Component {
 
     deleteMultipleRows = () => {
         let isOrg = delete_glossary.isOrg(this.props.glossaryData, this.state.rowsToDelete)
+        this.handleDeleteSelectedGlossaryBox()
         if (!isOrg) {
             let userId = JSON.parse(localStorage.getItem("userProfile")).userID
-            let rowsToBeDeleted = delete_glossary.getBulkDeletionArray(this.props.glossaryData, this.state.rowsToDelete)
-            this.makeDeleteGlossaryAPICall(userId, "", "", "", "", "JUDICIARY", true, rowsToBeDeleted)
+            let rowsToBeDeleted = delete_glossary.getBulkDeletionArray(this.props.glossaryData, this.state.rowsToDelete);
+            let orgID = JSON.parse(localStorage.getItem("userProfile")).orgID;
+            this.makeDeleteGlossaryAPICall(userId, orgID, "", "", "", "", "JUDICIARY", true, rowsToBeDeleted)
         } else {
             this.setState({ open: true, message: "Cannot delete glossary of type Organization..", variant: "error" })
         }
@@ -135,7 +263,8 @@ class MyGlossary extends React.Component {
                 options: {
                     filter: false,
                     sort: false,
-                    display: 'excluded'
+                    display: 'excluded',
+                    download: false
                 },
             },
             {
@@ -171,6 +300,7 @@ class MyGlossary extends React.Component {
                 options: {
                     sort: false,
                     empty: true,
+                    download: false,
                     customBodyRender: (value, tableMeta, updateValue) => {
                         if (tableMeta.rowData) {
                             return (
@@ -222,7 +352,7 @@ class MyGlossary extends React.Component {
                 this.setState({ rowsToDelete: allRowsSelected })
             },
             onRowsDelete: () => {
-                this.deleteMultipleRows()
+                this.setState({openDeleteSelectedGlossaryConfirmDialogue : true})
             }
         };
         return (
@@ -233,6 +363,8 @@ class MyGlossary extends React.Component {
                         <Spinner />
                         :
                         <MuiThemeProvider theme={getMuiTheme()}>
+                            {this.renderDeleteAllGlossaryButton()}
+                            {this.renderDeleteSelectedGlossaryConfirmBox()}
                             <MUIDataTable
                                 title={translate("common.page.title.glossary")}
                                 columns={columns}
